Move supervisor/ui events package
