+++ app/src/main/java/fr/acinq/phoenix/AppViewModel.kt
@@ -29,8 +29,8 @@ import fr.acinq.phoenix.background.EclairNodeService
 import fr.acinq.phoenix.background.ElectrumServer
 import fr.acinq.phoenix.background.KitState
 import fr.acinq.phoenix.db.PaymentMeta
-import fr.acinq.phoenix.events.PayToOpenNavigationEvent
-import fr.acinq.phoenix.events.RemovePendingSwapIn
+import fr.acinq.phoenix.background.PayToOpenNavigationEvent
+import fr.acinq.phoenix.background.RemovePendingSwapIn
 import fr.acinq.phoenix.utils.Constants
 import fr.acinq.phoenix.utils.ServiceDisconnected
 import fr.acinq.phoenix.utils.SingleLiveEvent
